Phase 4 commit 4: extract timeline/app.ts; main.ts becomes thin entry
createTimelineApp() owns everything timeline-related: data loading,
AppState construction, DOM lookups, render functions (renderTimeline,
renderSidebar), all interactions (wheel zoom, drag pan, ctrl-drag
reschedule, quick-add zone indicators, shift-preview, click handlers
for cards and toolbar buttons, session manager, advance time,
filter panel toggle, resize). Returns the high-level operations
(zoomBy / panBy / jumpToNow / collapseExpansion / openSearch /
isSearchOpen) that bootstrap/shortcuts.ts wires keyboard input to.

main.ts collapses to 27 lines:
  mountAppShell(appEl)
  → createViewSwitcher()
  → createTimelineApp()
  → attachGlobalShortcuts(timeline)
  → showNotes() if /notes path

No behaviour change. Build clean, tests baseline unchanged
(113/114 — only pre-existing calendar formatAxisDay failure).
